refactor: Refactor layout a + clean up unused files
