+++ components/layouts/SiteLayout.tsx
@@ -3,14 +3,14 @@ import Head from 'next/head';
 import Link from 'next/link';
 import * as React from 'react';
 
-export type DefaultLayoutProps = {
+export type SiteLayoutProps = {
    title: string;
 };
 
-export function DefaultLayout({
+export function SiteLayout({
    children,
    title,
-}: React.PropsWithChildren<DefaultLayoutProps>) {
+}: React.PropsWithChildren<SiteLayoutProps>) {
    return (
       <Box>
          <Head>
@@ -46,7 +46,13 @@ type IFixitLogoProps = {
 };
 
 const IFixitLogo = chakra(({ className }: IFixitLogoProps) => (
-   <svg width="118" height="36" fill="none" xmlns="http://www.w3.org/2000/svg">
+   <svg
+      className={className}
+      width="118"
+      height="36"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+   >
       <path
          d="M108.185 13.14h-3.6c-.949 0-1.727-.787-1.727-1.746 0-.958.778-1.745 1.727-1.745h10.945c.949 0 1.727.787 1.727 1.745 0 .96-.778 1.746-1.727 1.746h-3.6v11.974a1.869 1.869 0 01-1.873 1.894 1.868 1.868 0 01-1.872-1.894V13.14zM96.218 11.394c0-1.057.827-1.893 1.873-1.893 1.045 0 1.872.836 1.872 1.893v13.72a1.869 1.869 0 01-1.872 1.894 1.869 1.869 0 01-1.873-1.894v-13.72zM86.768 18.356l5.334-5.393a1.944 1.944 0 000-2.728 1.894 1.894 0 00-2.7 0l-5.334 5.392-5.334-5.392a1.894 1.894 0 00-2.699 0 1.944 1.944 0 000 2.728l5.334 5.393-5.334 5.392a1.944 1.944 0 000 2.728 1.892 1.892 0 002.7 0l5.334-5.392 5.334 5.392a1.892 1.892 0 002.699 0 1.945 1.945 0 000-2.728l-5.334-5.392zM68.174 11.394c0-1.057.827-1.893 1.872-1.893 1.046 0 1.873.836 1.873 1.893v13.72a1.869 1.869 0 01-1.873 1.894 1.868 1.868 0 01-1.872-1.894v-13.72zM52.047 11.542c0-1.057.828-1.893 1.874-1.893h9.534c.949 0 1.703.762 1.703 1.72 0 .96-.754 1.722-1.703 1.722h-7.662v3.663h6.568c.948 0 1.702.763 1.702 1.721a1.7 1.7 0 01-1.702 1.722h-6.568v4.917a1.869 1.869 0 01-1.872 1.894 1.869 1.869 0 01-1.874-1.894V11.542zM44.046 11.394c0-1.057.827-1.893 1.872-1.893 1.046 0 1.873.836 1.873 1.893v13.72a1.868 1.868 0 01-1.873 1.894 1.868 1.868 0 01-1.872-1.894v-13.72z"
          fill="#fff"
