+++ features/collection/components/CollectionBanner.tsx
