+++ features/collection/index.tsx
@@ -1,3 +1,3 @@
-export * from './views';
 export * from './loaders';
+export * from './components';
 export * from './types';
