--- features/collection/views/index.ts
@@ -1,2 +1,0 @@
-export * from './DefaultCollectionView';
-export * from './PartsCollectionView';
